custom_nodes/openai_image_edit/requirements.txt
openai>=1.0.0
pillow>=10.0.0
torch
numpy 